// src/common/enum/user.enum.ts
export enum GenderEnum {
    male = "male",
    female = "female"
}

export enum RoleEnum {
    user = "user",
    admin = "admin",
    superAdmin = "superAdmin"
}

export enum providerEnum {
    system = "system",
    google = "google",
    facebook = "facebook"
}

export enum EmailEnum {
    confirmedEmail = "confirmedEmail",
    forgetPassword = "forgetPassword"
}